Added creating analog output tags

+++ back/scada/scada/DTOS/AnalogOutputDTO.cs
@@ -6,11 +6,11 @@
         public string tagName { get; set; }
         public string Description { get; set; }
         public string IOAddress { get; set; }
-        public int InitialValue { get; set; }
+        public float InitialValue { get; set; }
         public double LowLimit { get; set; }
         public double HighLimit { get; set; }
         public string Units { get; set; }
-        public AnalogOutputDTO(float currentValue, string tagName, string description, string iOAddress, int initialValue, double lowLimit, double highLimit, string units)
+        public AnalogOutputDTO(float currentValue, string tagName, string description, string iOAddress, float initialValue, double lowLimit, double highLimit, string units)
         {
             this.currentValue = currentValue;
             this.tagName = tagName;
